Fix all lints

+++ quadratic-core/src/formulas/functions/mathematics.rs
@@ -441,8 +441,8 @@ mod tests {
         // Test mismatched range
         assert_eq!(
             RunErrorMsg::ExactArraySizeMismatch {
-                expected: ArraySize::try_from((1 as i64, 12 as i64)).unwrap(),
-                got: ArraySize::try_from((1 as i64, 11 as i64)).unwrap(),
+                expected: ArraySize::try_from((1_i64, 12_i64)).unwrap(),
+                got: ArraySize::try_from((1_i64, 11_i64)).unwrap(),
             },
             eval_to_err(&g, "SUMIFS(0..10, 0..11, \"<=5\")").msg,
         );
